+++ Services/ICallEdgarService.cs
@@ -4,6 +4,6 @@ namespace Fora.Services
 {
     public interface ICallEdgarService
     {
-        Task<EdgarCompanyInfo?> GetEdgarInfo(int cik);
+        Task<EdgarCompanyInfo?> GetEdgarInfo(long cik);
     }
 }
